Added order booking dropdowns, LCL Capacity

+++ lib/models/manage_orders/lcl/lcl_revise_allocation.dart
@@ -110,7 +110,7 @@ class Datum {
   dynamic destinationLocationBook;
   dynamic destDeliveryGatewayBook;
   dynamic isLoadingTallyEnabled;
-  int? decWeight;
+  double? decWeight;
   dynamic orderType;
   dynamic isReschedule;
   dynamic ffvId;
